chore: Only build one test file with integration tag
Build tags make LSPs unhappy. So instead, we'll just have it in one place
so the code is easier to edit.
