+++ cmd/unikraft/volumes_test.go
@@ -3,8 +3,6 @@
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
 
-//go:build integration
-
 package main
 
 var volumesTestCases = []testCase{
